+++ index.es6.exports/isPossibleNumber.js
@@ -1,3 +1,5 @@
+// Deprecated.
+
 import { withMetadata } from '../min/metadata'
 
 import _isPossibleNumber from '../es6/isPossibleNumber'