+++ min/exports/isValidPhoneNumber.js
@@ -1,0 +1,6 @@
+import { withMetadata } from '../metadata'
+import { isValidPhoneNumber as _isValidPhoneNumber } from '../../core/index'
+
+export function isValidPhoneNumber() {
+	return withMetadata(_isValidPhoneNumber, arguments)
+}