+++ min/index.js
@@ -4,6 +4,9 @@ export { parsePhoneNumberWithError, parsePhoneNumberWithError as parsePhoneNumbe
 // it has been promoted to a default export due to being too verbose.
 export { parsePhoneNumberFromString, parsePhoneNumberFromString as default } from './exports/parsePhoneNumberFromString'
 
+export { isValidPhoneNumber } from './exports/isValidPhoneNumber'
+export { isPossiblePhoneNumber } from './exports/isPossiblePhoneNumber'
+
 // Deprecated.
 export { findNumbers } from './exports/findNumbers'
 // Deprecated.